+++ src/AngleSharp.Io/Interfaces/IIndexedDbObjectStore.cs
@@ -1,0 +1,13 @@
+namespace AngleSharp.Io.Dom
+{
+    using AngleSharp.Attributes;
+
+    /// <summary>
+    /// Represents an IndexedDB object store.
+    /// </summary>
+    [DomName("IDBObjectStore")]
+    [DomExposed("Window")]
+    public interface IIndexedDbObjectStore : IStorage
+    {
+    }
+}