+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Released on Friday, July 31 2026.
 
 - Updated to use a minimum of AngleSharp 1.5
 - Fixed domain handling in cookies #36
-- Added `LocalStorage` and `SessionStorage`
+- Added `localStorage` and `sessionStorage` for web storages
+- Added `indexedDB` using an IndexedDB implementation
+- Added `caches` using a Cache API implementation
 
 # 1.0.0